making progress on lab 01
