+++ .gitattributes
@@ -1,0 +1,4 @@
+* text=auto
+
+# Align with eslint linebreak-style: [ "error", "unix" ]
+*.js text eol=lf
