+++ __tests__/arithmetic.test.js
@@ -1,5 +1,6 @@
 'use strict';
 
 const arithmetic = require('../lib/arithmetic.js');
-const faker = require('faker');
+
+describe('Arithmetic Module - Addition', () => {
